maybe he wants to be a float. /google..x2

+++ task_03.py
@@ -54,7 +54,7 @@ elif LOAN >= 1000000:
 N = 12
 if RATE is not None:
     RATE = decimal.Decimal(RATE)
-    TOTAL = float(round(LOAN * (1 + RATE/N)**(N*YEARS)))
+    TOTAL = int(round(LOAN * (1 + RATE/N)**(N*YEARS)))
 else:
     TOTAL = None
